+++ apps/wallet-mobile/src/features/Send/useCases/ConfirmTx/FailedTx/FailedTxImage.tsx
