+++ apps/wallet-mobile/src/features/Send/useCases/ConfirmTx/SubmittedTx/SubmittedTxScreen.tsx
@@ -1,10 +1,10 @@
 import React from 'react'
 import {StyleSheet, View} from 'react-native'
 
-import {Button, Spacer, Text} from '../../../../components'
-import {useWalletNavigation} from '../../../../navigation'
-import {COLORS} from '../../../../theme'
-import {useStrings} from '../../common/strings'
+import {Button, Spacer, Text} from '../../../../../components'
+import {useWalletNavigation} from '../../../../../navigation'
+import {COLORS} from '../../../../../theme'
+import {useStrings} from '../../../common/strings'
 import {SubmittedTxImage} from './SubmittedTxImage'
 
 export const SubmittedTxScreen = () => {
