+++ apps/wallet-mobile/translations/messages/src/AppNavigator.json
