+++ apps/wallet-mobile/translations/messages/src/Dashboard/WithdrawStakingRewards/WithdrawStakingRewards.json
