+++ apps/wallet-mobile/translations/messages/src/TxHistory/TxHistoryNavigator.json
@@ -4,14 +4,14 @@
     "defaultMessage": "!!!Receive",
     "file": "src/TxHistory/TxHistoryNavigator.tsx",
     "start": {
-      "line": 162,
+      "line": 168,
       "column": 16,
-      "index": 4973
+      "index": 5382
     },
     "end": {
-      "line": 165,
+      "line": 171,
       "column": 3,
-      "index": 5062
+      "index": 5471
     }
   },
   {
@@ -19,14 +19,14 @@
     "defaultMessage": "!!!Send",
     "file": "src/TxHistory/TxHistoryNavigator.tsx",
     "start": {
-      "line": 166,
+      "line": 172,
       "column": 13,
-      "index": 5077
+      "index": 5486
     },
     "end": {
-      "line": 169,
+      "line": 175,
       "column": 3,
-      "index": 5157
+      "index": 5566
     }
   },
   {
@@ -34,14 +34,14 @@
     "defaultMessage": "!!!Scan QR code address",
     "file": "src/TxHistory/TxHistoryNavigator.tsx",
     "start": {
-      "line": 170,
+      "line": 176,
       "column": 18,
-      "index": 5177
+      "index": 5586
     },
     "end": {
-      "line": 173,
+      "line": 179,
       "column": 3,
-      "index": 5278
+      "index": 5687
     }
   },
   {
@@ -49,14 +49,14 @@
     "defaultMessage": "!!!Select asset",
     "file": "src/TxHistory/TxHistoryNavigator.tsx",
     "start": {
-      "line": 174,
+      "line": 180,
       "column": 20,
-      "index": 5300
+      "index": 5709
     },
     "end": {
-      "line": 177,
+      "line": 183,
       "column": 3,
-      "index": 5389
+      "index": 5798
     }
   },
   {
@@ -64,14 +64,14 @@
     "defaultMessage": "!!!Selected tokens",
     "file": "src/TxHistory/TxHistoryNavigator.tsx",
     "start": {
-      "line": 178,
+      "line": 184,
       "column": 26,
-      "index": 5417
+      "index": 5826
     },
     "end": {
-      "line": 181,
+      "line": 187,
       "column": 3,
-      "index": 5521
+      "index": 5930
     }
   },
   {
@@ -79,14 +79,14 @@
     "defaultMessage": "!!!Edit amount",
     "file": "src/TxHistory/TxHistoryNavigator.tsx",
     "start": {
-      "line": 182,
+      "line": 188,
       "column": 19,
-      "index": 5542
+      "index": 5951
     },
     "end": {
-      "line": 185,
+      "line": 191,
       "column": 3,
-      "index": 5635
+      "index": 6044
     }
   },
   {
@@ -94,14 +94,14 @@
     "defaultMessage": "!!!Confirm",
     "file": "src/TxHistory/TxHistoryNavigator.tsx",
     "start": {
-      "line": 186,
+      "line": 192,
       "column": 16,
-      "index": 5653
+      "index": 6062
     },
     "end": {
-      "line": 189,
+      "line": 195,
       "column": 3,
-      "index": 5739
+      "index": 6148
     }
   },
   {
@@ -109,14 +109,14 @@
     "defaultMessage": "!!!Share this address to receive payments. To protect your privacy, new addresses are generated automatically once you use them.",
     "file": "src/TxHistory/TxHistoryNavigator.tsx",
     "start": {
-      "line": 190,
+      "line": 196,
       "column": 19,
-      "index": 5760
+      "index": 6169
     },
     "end": {
-      "line": 196,
+      "line": 202,
       "column": 3,
-      "index": 5998
+      "index": 6407
     }
   }
 ]
