+++ apps/wallet-mobile/translations/messages/src/WalletInit/MnemonicExplanationModal/MnemonicExplanationModal.json
