+++ apps/wallet-mobile/translations/messages/src/components/LanguagePicker/LanguagePickerWarning.json
